+++ R/crs-funs.R
@@ -13,7 +13,9 @@
 #' @export
 #' @examples
 #' data("routes_fast")
-#' new_crs <- crs_select_aeq(routes_fast)
+#' new_crs <- geo_select_aeq(routes_fast)
+#' new_crs2 <- geo_select_aeq(routes_fast) # to be deprecated
+#' identical(new_crs, new_crs2)
 #' plot(routes_fast)
 #' rf_projected <- sp::spTransform(routes_fast, new_crs)
 #' plot(rf_projected)
@@ -22,6 +24,7 @@
 #' plot(line_length, rf_projected$length)
 #' cor(line_length, rf_projected$length)
 crs_select_aeq <- function(shp) {
+  .Deprecated(new = "geo_select_aeq")
   cent <- rgeos::gCentroid(shp)
   aeqd <- sprintf(
     "+proj=aeqd +lat_0=%s +lon_0=%s +x_0=0 +y_0=0",
@@ -35,14 +38,14 @@ crs_select_aeq <- function(shp) {
 #'
 #' @param shp A spatial object with a geographic (WGS84) coordinate system
 #' @param crs An optional coordinate reference system (if not provided it is set
-#' automatically by [crs_select_aeq()]).
+#' automatically by [geo_select_aeq()]).
 #' @family geo
 #' @export
 #' @examples
 #' data(routes_fast)
 #' rf_aeq <- reproject(routes_fast[1:3, ])
 #' rf_osgb <- reproject(routes_fast[1:3, ], 27700)
-reproject <- function(shp, crs = crs_select_aeq(shp)) {
+reproject <- function(shp, crs = geo_select_aeq(shp)) {
   if (is.na(raster::crs(shp))) {
     message("Assuming a geographical (lat/lon) CRS (EPSG:4326)")
     raster::crs(shp) <- sp::CRS("+init=epsg:4326")